Move jasonifying util function to service.
