+++ python/lsst/ts/logging_and_reporting/web_app/main.py
@@ -6,7 +6,7 @@ from fastapi.responses import JSONResponse
 from fastapi.encoders import jsonable_encoder
 from lsst.ts.logging_and_reporting.exceptions import ConsdbQueryError, BaseLogrepError
 from lsst.ts.logging_and_reporting.utils import get_access_token
-import numpy as np
+# import numpy as np
 
 from .services.jira_service import get_jira_tickets
 from .services.consdb_service import get_mock_exposures, get_exposures, get_data_log
@@ -90,19 +90,6 @@ async def read_exposures(
         raise HTTPException(status_code=500, detail=str(e))
 
 
-# TODO: Confirm where this util func should live.
-# Required to jasonify pandas DataFrame with special float values
-def stringify_special_floats(val):
-    if isinstance(val, float):
-        if np.isnan(val):
-            return "NaN"
-        elif np.isposinf(val):
-            return "Infinity"
-        elif np.isneginf(val):
-            return "-Infinity"
-    return val
-
-
 @app.get("/data-log")
 async def read_data_log(
     request: Request,
@@ -115,16 +102,7 @@ async def read_data_log(
     try:
         auth_header = request.headers.get("Authorization")
         auth_token = auth_header.split(" ")[1] if auth_header else None
-
-        # Returns a pandas DataFrame
-        df = get_data_log(dayObsStart, dayObsEnd, instrument, auth_token)
-
-        # Convert special floats (nans and infs) to strings
-        # This ensures that JSON serialisation does not fail
-        logger.debug(f"Converting DataFrame to JSON records: {df.shape}")
-        df_safe = df.applymap(stringify_special_floats)
-        records = df_safe.to_dict(orient="records")
-
+        records = get_data_log(dayObsStart, dayObsEnd, instrument, auth_token)
         return jsonable_encoder({"data_log": records})
 
     except ConsdbQueryError as ce:
